# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finincial accounting API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/earnings/{id}": {
            "get": {
                "tags": [
                    "Earnings"
                ],
                "summary": "Get earning by id",
                "description": "Get earning by id",
                "operationId": "b09470a7be14a4dc66f5cef317e2ca30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Earning id",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Earnings"
                ],
                "summary": "Delete Earning",
                "description": "Delete earning by ID",
                "operationId": "earningDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Earning ID",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Earnings"
                ],
                "summary": "Update Earning",
                "description": "Update Earning by ID",
                "operationId": "earningUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Earning ID",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": "100"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-28"
                                    },
                                    "source_id": {
                                        "type": "id",
                                        "example": "1"
                                    },
                                    "type_id": {
                                        "type": "id",
                                        "example": "1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Example description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Updated Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/earnings": {
            "post": {
                "tags": [
                    "Earnings"
                ],
                "summary": "Create Earning",
                "description": "Create Earning",
                "operationId": "earningCreate",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "date",
                                    "source_id",
                                    "type_id"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": "100"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-28"
                                    },
                                    "source_id": {
                                        "type": "id",
                                        "example": "1"
                                    },
                                    "type_id": {
                                        "type": "id",
                                        "example": "1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Example description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/spednings/{id}": {
            "delete": {
                "tags": [
                    "Spendings"
                ],
                "summary": "Delete Spending",
                "description": "Delete spedning by ID",
                "operationId": "spedningDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Spending ID",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
        "/spendings/{id}": {
            "get": {
                "tags": [
                    "Spendings"
                ],
                "summary": "Get spending by id",
                "description": "Get spending by id",
                "operationId": "ea12106d237d98fc92faf3433dc6de85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Spending id",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Spendings"
                ],
                "summary": "Update Spending",
                "description": "Update Spending by ID",
                "operationId": "spendingUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Spending ID",
                        "required": true,
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": "100"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-28"
                                    },
                                    "category_id": {
                                        "type": "id",
                                        "example": "1"
                                    },
                                    "type_id": {
                                        "type": "id",
                                        "example": "1"
                                    },
                                    "tag_ids": {
                                        "type": "array",
                                        "items": {},
                                        "example": "[\"McDonalds\", \"BurgerKing\", \"KFC\"]"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Example description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Updated Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}